Extract SettingsView
