+++ VoiceOver Designer/Features/Sources/Settings/States/Control/SettingsViewController.swift
@@ -8,10 +8,6 @@
 import AppKit
 import Document
 
-class TraitCheckBox: NSButton {
-    var trait: A11yTraits!
-}
-
 public class SettingsViewController: NSViewController {
 
     public var presenter: SettingsPresenter!
@@ -20,93 +16,17 @@ public class SettingsViewController: NSViewController {
         presenter.model
     }
     
-    @IBOutlet weak var resultLabel: NSTextField!
-    
-    @IBOutlet weak var scrollView: NSScrollView!
-    @IBOutlet weak var mainStack: NSStackView!
-    @IBOutlet weak var label: NSTextField!
-    @IBOutlet weak var hint: NSTextField!
-    
-    // MARK: Type trait
-    @IBOutlet weak var buttonTrait: TraitCheckBox!
-    @IBOutlet weak var headerTrait: TraitCheckBox!
-
-    @IBOutlet weak var linkTrait: TraitCheckBox!
-    @IBOutlet weak var staticTextTrait: TraitCheckBox!
-    @IBOutlet weak var imageTrait: TraitCheckBox!
-    @IBOutlet weak var searchFieldTrait: TraitCheckBox!
-    @IBOutlet weak var tabTrait: TraitCheckBox!
-    
-    @IBOutlet weak var selectedTrait: TraitCheckBox!
-    @IBOutlet weak var summaryElementTrait: TraitCheckBox!
-    @IBOutlet weak var playSoundTrait: TraitCheckBox!
-    @IBOutlet weak var allowsDirectInteraction: TraitCheckBox!
-    @IBOutlet weak var startMediaSession: TraitCheckBox!
-    @IBOutlet weak var disabledTrait: TraitCheckBox!
-    @IBOutlet weak var updatesFrequently: TraitCheckBox!
-    @IBOutlet weak var causesPageTurn: TraitCheckBox!
-    @IBOutlet weak var keyboardKey: TraitCheckBox!
-    
-    @IBOutlet weak var isAccessibilityElement: NSButton!
-    
     // MARK: behaviourTrait
     // TODO: Add
     
     public override func viewDidLoad() {
         super.viewDidLoad()
         
-        scrollView.documentView = mainStack
-        setup(from: descr)
+        view().setup(from: descr)
     }
     
-    private func setup(from descr: A11yDescription) {
-        label.stringValue = descr.label
-        hint.stringValue  = descr.hint
-        isAccessibilityElement.state = descr.isAccessibilityElement ? .on: .off
-        
-        updateText(isUserAction: false)
-        
-        buttonTrait.trait = .button
-        headerTrait.trait = .header
-        linkTrait.trait = .link
-        staticTextTrait.trait = .staticText
-        imageTrait.trait = .image
-        searchFieldTrait.trait = .searchField
-        tabTrait.trait = .tab
-        
-        selectedTrait.trait = .selected
-        summaryElementTrait.trait = .summaryElement
-        playSoundTrait.trait = .playsSound
-        allowsDirectInteraction.trait = .allowsDirectInteraction
-        startMediaSession.trait = .startsMediaSession
-        disabledTrait.trait = .notEnabled
-        updatesFrequently.trait = .updatesFrequently
-        causesPageTurn.trait = .causesPageTurn
-        keyboardKey.trait = .keyboardKey
-        
-        let allTraitsButtons: [TraitCheckBox] = [
-            buttonTrait,
-            headerTrait,
-            linkTrait,
-            staticTextTrait,
-            imageTrait,
-            searchFieldTrait,
-            tabTrait,
-            
-            selectedTrait,
-            summaryElementTrait,
-            playSoundTrait,
-            allowsDirectInteraction,
-            startMediaSession,
-            disabledTrait,
-            updatesFrequently,
-            causesPageTurn,
-            keyboardKey,
-        ]
-        
-        for traitButton in allTraitsButtons {
-            traitButton.state = descr.trait.contains(traitButton.trait) ? .on: .off
-        }
+    func view() -> SettingsView {
+        view as! SettingsView
     }
     
     weak var valueViewController: A11yValueViewController?
@@ -136,27 +56,25 @@ public class SettingsViewController: NSViewController {
         } else {
             descr.trait.subtract(sender.trait)
         }
-        updateText(isUserAction: true)
+        updateText()
     }
     
     // MARK: Description
     @IBAction func labelDidChange(_ sender: NSTextField) {
         // TODO: if you forgot to call updateColor, the label wouldn't be revalidated
         presenter.updateLabel(to: sender.stringValue)
-        updateText(isUserAction: true)
+        updateText()
     }
     
     @IBAction func hintDidChange(_ sender: NSTextField) {
         descr.hint = sender.stringValue
-        updateText(isUserAction: true)
+        updateText()
     }
     
-    func updateText(isUserAction: Bool) {
-        resultLabel.stringValue = descr.voiceOverText
+    func updateText() {
+        view().updateText(from: descr)
         
-        if isUserAction {
-            presenter.delegate?.didUpdateValue()
-        }
+        presenter.delegate?.didUpdateValue()
     }
     
     @IBAction func delete(_ sender: Any) {
