+++ VoiceOver Designer/Features/Sources/Settings/States/Control/Value/A11yValueViewController.swift
@@ -9,7 +9,7 @@ import AppKit
 import Document
 
 protocol A11yValueDelegate: AnyObject {
-    func updateText(isUserAction: Bool)
+    func updateText()
 }
 
 class A11yValueViewController: NSViewController {
@@ -40,7 +40,7 @@ class A11yValueViewController: NSViewController {
     
     @IBAction func valueDidChange(_ sender: NSTextField) {
         descr.value = sender.stringValue
-        delegate?.updateText(isUserAction: true)
+        delegate?.updateText()
     }
     
     @IBAction func addAdjustable(_ sender: Any) {
@@ -50,7 +50,7 @@ class A11yValueViewController: NSViewController {
         
         view().selectLastOption()
         
-        delegate?.updateText(isUserAction: true)
+        delegate?.updateText()
     }
     
     func saveCurrentChanges() {
@@ -69,13 +69,13 @@ class A11yValueViewController: NSViewController {
         }
         
         renderDescription(setFirstResponder: false)
-        delegate?.updateText(isUserAction: true)
+        delegate?.updateText()
     }
     
     @IBAction func isEnumeratedDidChanged(_ sender: NSButton) {
         descr.isEnumeratedAdjustable = sender.state == .on
         
-        delegate?.updateText(isUserAction: true)
+        delegate?.updateText()
     }
     
     func renderDescription(setFirstResponder: Bool) {
@@ -103,7 +103,7 @@ extension A11yValueViewController: AdjustableOptionViewDelegate {
         
         view().deselectRadioGroup(selected: option)
         
-        delegate?.updateText(isUserAction: true)
+        delegate?.updateText()
     }
     
     func update(option: AdjustableOptionView) {
@@ -112,6 +112,6 @@ extension A11yValueViewController: AdjustableOptionViewDelegate {
                                            with: option.text)
         }
         
-        delegate?.updateText(isUserAction: true)
+        delegate?.updateText()
     }
 }
